Handle integer pixel scaling as well

+++ layer/TimeValueLayer.cpp
@@ -565,7 +565,7 @@ TimeValueLayer::getLocalPoints(LayerGeometryProvider *v, int x) const
     }
 
     if (!usePoints.empty()) {
-        int fuzz = v->scaleSize(2);
+        double fuzz = v->scaleSize(2);
         int px = v->getXForFrame(usePoints.begin()->frame);
         if ((px > x && px - x > fuzz) ||
             (px < x && x - px > fuzz + 3)) {
